requirements.txt
-i https://pypi.org/simple
aiohttp==3.8.5
aiosignal==1.3.1 ; python_version >= '3.7'
aiosqlite==0.17.0 ; python_version >= '3.6'
async-timeout==4.0.3 ; python_version >= '3.7'
attrs==23.1.0 ; python_version >= '3.7'
beautifulsoup4==4.12.2 ; python_full_version >= '3.6.0'
brotli==1.0.9 ; platform_python_implementation == 'CPython'
bs4==0.0.1
certifi==2023.7.22 ; python_version >= '3.6'
charset-normalizer==3.2.0 ; python_full_version >= '3.7.0'
click==8.1.7 ; python_version >= '3.7'
frozenlist==1.4.0 ; python_version >= '3.8'
hiredis==2.2.3
idna==3.4 ; python_version >= '3.5'
iniconfig==2.0.0 ; python_version >= '3.7'
iso8601==1.1.0 ; python_version < '4.0' and python_full_version >= '3.6.2'
lxml==4.9.3 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
multidict==6.0.4 ; python_version >= '3.7'
mutagen==1.47.0 ; python_version >= '3.7'
packaging==23.1 ; python_version >= '3.7'
pluggy==1.3.0 ; python_version >= '3.8'
pycryptodomex==3.18.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
pydantic==1.10.11
pypika-tortoise==0.1.6 ; python_version >= '3.7' and python_version < '4.0'
pytest==7.4.1 ; python_version >= '3.7'
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
redis[hiredis]==5.0.0
requests==2.31.0 ; python_version >= '3.7'
rq==1.15.1
rss-parser==0.2.4
six==1.16.0 ; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
soupsieve==2.5 ; python_version >= '3.8'
tortoise-orm==0.20.0
typing-extensions==4.7.1 ; python_version >= '3.7'
urllib3==2.0.4 ; python_version >= '3.7'
websockets==11.0.3 ; python_version >= '3.7'
yarl==1.9.2 ; python_version >= '3.7'
yt-dlp==2023.7.6
